ATR-762: moved unit tests for PX1094 to DAC and DAC extension declaration tests
